Remove the loadout/build tab, move contents to header modals
