+++ libs/gi/localization/assets/locales/en/page_character.json
@@ -1,7 +1,6 @@
 {
   "addNew": "Add Character",
   "tabs": {
-    "setting": "Loadout/Build",
     "overview": "Overview",
     "talent": "Talents",
     "optimize": "Optimize",
